% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beliaeff2002.R
\name{beliaeff2002}
\alias{beliaeff2002}
\title{Compute Integrated Biomarker Index}
\usage{
beliaeff2002(sam_mean)
}
\arguments{
\item{sam_mean}{a data frame, the third output of \code{\link{simul}}.}
}
\value{
\code{beliaeff2002} returns a data frame of IBR.
}
\description{
\code{beliaeff2002} calculates IBR in the hypothetical case study. This is
not meant to be called directly.
}
\references{
Beliaeff, B., & Burgeot, T. (2002). Integrated biomarker
  response: A useful tool for ecological risk assessment. Environmental
  Toxicology and Chemistry, 21(6), 1316–1322. \doi{10.1002/etc.5620210629}.
}
