% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con2ord.R
\name{con2ord}
\alias{con2ord}
\title{Ordinalize Biomarker Ordinal Estimator}
\usage{
con2ord(x, thresholds = c(0.25, 0.5, 0.75))
}
\arguments{
\item{x}{a vector, the BOE values.}

\item{thresholds}{a numeric vector of length 3, the cut points for
ordinalization. The default values are 0.25, 0.5, and 0.75.}
}
\value{
\code{con2ord} returns a factor with 5 levels (i.e., no, very small,
  small, medium, and large) that are the perceived differences from BOE.
}
\description{
\code{con2ord} ordinalizes the Biomarker Ordinal Estimator (BOE) into very
small, small, medium, and large differences.
}
\examples{
con2ord(seq(0, 1, 0.1))

}
