% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbar.R
\name{ggbar}
\alias{ggbar}
\title{Make Bar Plot}
\usage{
ggbar(dat, hax, vax, sub, env = parent.frame())
}
\arguments{
\item{dat}{a data frame with at least two columns.}

\item{hax}{a character, name of the column to be used as the horizontal axis.}

\item{vax}{a character, name of the column to be used as the vertical axis.}

\item{sub}{a numeric, mean ecological relevance.}

\item{env}{an environment, to access outer scope variables.}
}
\value{
\code{ggbar} returns a ggplot object.
}
\description{
\code{ggbar} creates bar plot of the ecological relevance. This is not meant
to be called directly.
}
