% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdot.R
\name{ggdot}
\alias{ggdot}
\title{Make Dot Plot}
\usage{
ggdot(dat, hax, vax)
}
\arguments{
\item{dat}{a data frame with at least two columns.}

\item{hax}{a character, name of the column to be used as the horizontal axis.}

\item{vax}{a character, name of the column to be used as the vertical axis.}
}
\value{
\code{ggdot} returns a ggplot object.
}
\description{
\code{ggdot} creates dot plot of the average of absolute Cliff’s delta. This
is not meant to be called directly.
}
