% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{Make Heatmap}
\usage{
ggheat(
  dat,
  hax,
  vax,
  cell,
  nm,
  lim,
  lo,
  hi,
  diverging = FALSE,
  env = parent.frame()
)
}
\arguments{
\item{dat}{a data frame with at least three columns.}

\item{hax}{a character, name of the column to be used as the horizontal axis.}

\item{vax}{a character, name of the column to be used as the vertical axis.}

\item{cell}{a character, name of the column to be used as the cells.}

\item{nm}{a character, name of the heatmap.}

\item{lim}{a numeric vector, limits of the color scale.}

\item{lo}{a character, color of the color scale low end.}

\item{hi}{a character, color of the color scale high end.}

\item{diverging}{a logical, whether to use diverging color gradient.}

\item{env}{an environment, to access outer scope variables.}
}
\value{
\code{ggheat} returns a ggplot object.
}
\description{
\code{ggheat} creates heatmaps of the Cliff's delta and S-value. This is not
meant to be called directly.
}
