% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{heat}
\alias{heat}
\title{Visualize Biomarker Ordinal Estimator}
\usage{
heat(dat, hax, vax, cell)
}
\arguments{
\item{dat}{a data frame with at least three columns.}

\item{hax}{a character, name of the column to be used as the horizontal axis.}

\item{vax}{a character, name of the column to be used as the vertical axis.}

\item{cell}{a character, name of the column to be used as the cells.}
}
\value{
\code{heat} returns a ggplot object.
}
\description{
\code{heat} plots a heatmap of the perceived differences from Biomarker
Ordinal Estimator (BOE).
}
