% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr.cliff.R
\name{mbr.cliff}
\alias{mbr.cliff}
\title{Compute Cliff's delta simplified}
\usage{
mbr.cliff(df)
}
\arguments{
\item{df}{a data frame with the name of experimental groups or biomonitoring
sites as the first column and the measurement of biomarkers as the
remaining columns.}
}
\description{
\code{mbr.cliff} summarizes Cliff's delta for multiple groups and
multiple biomarkers.
}
\examples{

\donttest{
set.seed(1)
setting <- setpop()
temp <- simul(setting$pop_mean)
mbr.cliff(temp$sam)
} #might take more than 5s in some machines

}
