% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr.glass.R
\name{mbr.glass}
\alias{mbr.glass}
\title{Compute Glass's delta simplified}
\usage{
mbr.glass(df)
}
\arguments{
\item{df}{a data frame with the name of experimental groups or biomonitoring
sites as the first column and the measurement of biomarkers as the
remaining columns.}
}
\description{
\code{mbr.glass} summarizes Glass's delta for multiple groups and
multiple biomarkers.
}
\examples{

\donttest{
set.seed(1)
setting <- setpop()
temp <- simul(setting$pop_mean)
mbr.glass(temp$sam)
} #might take more than 5s in some machines

}
