% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbRes-package.R
\docType{package}
\name{mbRes-package}
\alias{mbRes-package}
\title{mbRes: Integrating Multiple Biomarker Responses in Aquatic Organisms using
Effect Size, Statistical Uncertainty, and Ecological Relevance}
\description{
Compute and visualize the ps-index, a new integrated index for multiple
biomarker responses, as described in Pham & Sokolova (2021, unpublished).
}
\section{Guidelines}{
 ps-index is a new integrated index for multiple
  biomarker responses. \code{\link{mbr}} is the main function to compute and
  visualize the ps-index. \code{\link{sokolova2021}} is provided as a sample
  dataset. \code{\link{compare}} simulates a hypothetical dataset and compare
  the results of ps-index and two other integrated indices published earlier.
  The others are helper functions and are not meant to be called directly by
  users.
}

\section{Copyright}{
 mbRes: Integrating Multiple Biomarker Responses in
  Aquatic Organisms using Effect Size, Statistical Uncertainty, and
  Ecological Relevance. Copyright (C) 2021  Duy Nghia Pham & Inna M. Sokolova
  \cr \cr mbRes is free software: you can redistribute it and/or modify it
  under the terms of the GNU General Public License as published by the Free
  Software Foundation, either version 3 of the License, or (at your option)
  any later version. \cr \cr mbRes is distributed in the hope that it will be
  useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
  Public License for more details. \cr \cr You should have received a copy of
  the GNU General Public License along with mbRes.  If not, see
  \url{https://www.gnu.org/licenses/}.
}

\author{
Duy Nghia Pham & Inna M. Sokolova
}
