% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perch.R
\name{perch}
\alias{perch}
\title{Empirical Case Study Analysis}
\usage{
perch()
}
\value{
\code{perch} returns a list of length 3: \item{\code{tab}}{a list of
  length 2 \describe{ \item{\code{hansson2014}}{biomarker data reported by
  Hansson et al. (2014)} \item{\code{percheco}}{assigned ecological relevance
  of biomarkers}
  }} \item{\code{fig}}{a list of 5 ggplot objects \describe{
  \item{\code{SG}}{length-corrected somatic growth} \item{\code{SCI}}{somatic
  condition index} \item{\code{GSI}}{time-corrected gonadosomatic index}
  \item{\code{LSI}}{time-corrected liver somatic index}
  \item{\code{EROD}}{liver ethoxyresorufin-O-deethylase activity} } }
  \item{\code{est}}{full results of our estimation method given by
  \code{mbr}}
}
\description{
\code{perch} analyses the biomarker results of Hansson et al. (2014)
<doi:10.1007/s00244-013-9974-5>.
}
\references{
Hansson, T., Hansen, W., Tjärnlund, U., Balk, L., & Bengtsson,
  B.-E. (2014). Biomarker Investigations in Adult Female Perch (Perca
  fluviatilis) From Industrialised Areas in Northern Sweden in 2003. Archives
  of Environmental Contamination and Toxicology, 66(2), 237–247.
  \doi{10.1007/s00244-013-9974-5}.
}
