% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpop.R
\name{setpop}
\alias{setpop}
\title{Define Hypothetical Populations}
\usage{
setpop()
}
\value{
\code{setpop} returns a list of length 3:
\item{\code{pop_mean}}{true means of biomarker responses in populations.}
\item{\code{pop_mean_long}}{true means of biomarker responses in long format.}
\item{\code{pop_profile}}{profile of biomarkers.}
}
\description{
\code{setpop} sets the true means of biomarker responses in populations. This
is used for the hypothetical case study.
}
