% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{simul}
\alias{simul}
\title{Generate Hypothetical Samples}
\usage{
simul(pop_mean, size = 75)
}
\arguments{
\item{pop_mean}{a data frame, the first output of \code{setpop}.}

\item{size}{an integer, the sample size.}
}
\value{
\code{simul} returns a list of length 3:
\item{\code{sam}}{sample dataset.}
\item{\code{sam_long}}{sample dataset in long format.}
\item{\code{sam_mean}}{sample means of biomarker responses.}
}
\description{
\code{simul} yields a sample dataset of biomarker responses. This
is used for the hypothetical case study.
}
