% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Hypothetical Case Study Analysis}
\usage{
compare()
}
\value{
\code{compare} returns a list of length 3: \item{\code{tab}}{a list
  of length 9 \describe{ \item{\code{pop}}{the true population means of
  normal distributions} \item{\code{pop_long}}{the true population means in
  long format} \item{\code{sam}}{the sample dataset of biomarker responses}
  \item{\code{sam_long}}{the sample dataset in long format}
  \item{\code{m}}{mean biomarker responses by exposure group}
  \item{\code{blaise2002}}{integrated index from an ordinalization method}
  \item{\code{beliaeff2002}}{integrated index from a standardization method}
  \item{\code{eco}}{assigned ecological relevance of biomarkers}
  \item{\code{pham2021}}{ps-index from our estimation method} }}
  \item{\code{fig}}{a list of 2 ggplot objects \describe{
  \item{\code{fig_sam}}{hypothetical dataset}
  \item{\code{fig_compare}}{result comparision of three methods} } }
  \item{\code{est}}{full results of our estimation method given by
  \code{mbr}}
}
\description{
\code{compare} generates a hypothetical dataset of biomarker responses and
analyzes the dataset using three different strategies.
}
